Added TabBarView to RootView
